factor out cycledash.tasks
